Move the Vault client code into the internal package
