+++ src/main/java/com/avioconsulting/mule/connector/vault/provider/internal/vault/client/VaultConstants.java
@@ -1,4 +1,4 @@
-package com.avioconsulting.mule.vault.api.client;
+package com.avioconsulting.mule.connector.vault.provider.internal.vault.client;
 
 public final class VaultConstants {
