+++ src/main/java/com/avioconsulting/mule/connector/vault/provider/internal/vault/client/exception/AccessException.java
@@ -1,4 +1,4 @@
-package com.avioconsulting.mule.vault.api.client.exception;
+package com.avioconsulting.mule.connector.vault.provider.internal.vault.client.exception;
 
 public class AccessException extends Exception {
     public AccessException(String errorMessage) {
